Added FpsLimiter and InputManager
FpsLimiter class does calculate fps and limits the fps to certain value

InputManger adds currently pressed keys to an unordered map, game logic
will react according to which keysare currently pressed.

Some changes to window class so that VYSNC mode can be changed in the
MainGame create init as all other revelant userrelated features are set
there. this makes it easier for the user to controll and change
important setings.