Added a world and object polymorphism
Added a world class that handles the move detection, later physics may
be added. Not sure if it wont be added to the Engine.

Every moving object is now chils of the new Object class. It holds key
movement information.

MainGame handles object movement of any kind unified. A template
function recieves an object(child of Objects). Movement is handled by
this 1 function. As it uses a tmplate, typeSafety is no longer given.
This shouldnt cause any problems, as everything it calls is only defined
once, wether everything or nothing works.