Added Collision handling and randomEngine
Collsion has still a lot of bugs.
Workflow: The physicsmanger works as connection piece between, the game
and the physicsEngine,

Collision with a static object id reflected in the same direction. I
plan on solving this with normals.

The randomengine creates an engine with every call, this is bad but
creating it on objectcreation didnt work.